refactor(schema/user-backend/controller): refactored errors and moved db calls into transactions and added patch validation
