+++ backend/controller/controller.go
@@ -27,8 +27,6 @@ func RegisterHandlers(e *echo.Echo, dbClient *mongo.Client, logger *pterm.Logger
 	// Register extra route handlers
 	RegisterAuthHandlers(api, store, dbClient, logger)
 	RegisterDesignHandlers(api, dbClient, logger)
-
-	// register all handlers
 	RegisterUserHandlers(api, dbClient, logger)
 
 	// catch any invalid endpoints with a 404 error
